Align ScaleTypeChoiceGroup with base props
