+++ packages/fluent/src/components/ScaleTypeChoiceGroup.types.ts
@@ -2,13 +2,8 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import type { IChoiceGroupStyles } from '@fluentui/react'
-import type { ScaleType } from '@thematic/core'
+import type { IChoiceGroupProps } from '@fluentui/react'
 
-export interface ScaleTypeChoiceGroupProps {
-	selectedType: ScaleType
-	onChange?: (scaleType: ScaleType) => void
-	label: string
+export interface ScaleTypeChoiceGroupProps extends IChoiceGroupProps {
 	suppressQuantile?: boolean
-	styles?: IChoiceGroupStyles
 }
